Add up to 4 ICEs (#360)
* Add two or three ICEs

* Make 71506 use fewer relative paths

* Add up to four more ICEs

66922 might need an MVCE to avoid the git commands

* Add four more ICEs

* Revise 71506 further

71506 now has only one mkdir command.

* Remove 71639 duplicate

I erroneously thought this was fixed, but it turns out that I had made
the files in the wrong places in the shell script.

* Restore 71922.rs

Side note: can someone teach me on how to resolve merge conflicts
properly?

* Remove excess temp directory boilerplate

Also, 71506 is now cleaned up and clarified.

* Remove reliance on git clones from github

This will definitely speed things up. 66922 is big, but I bet it can be
pared down.

* Remove six ICEs

Per @JohnTitor's suggestion to make compile time not slower than it
already is.