ices/83471.rs: fixed with errors
=== stdout ===
=== stderr ===
error: expected one of `:`, `@`, or `|`, found `)`
 --> /home/runner/work/glacier/glacier/ices/83471.rs:3:24
  |
3 |     fn call(export_name);
  |                        ^ expected one of `:`, `@`, or `|`
  |
  = note: anonymous parameters are removed in the 2018 edition (see RFC 1685)
help: if this is a `self` type, give it a parameter name
  |
3 |     fn call(self: export_name);
  |             ~~~~~~~~~~~~~~~~~
help: if this is a parameter name, give it a type
  |
3 |     fn call(export_name: TypeName);
  |             ~~~~~~~~~~~~~~~~~~~~~
help: if this is a type, explicitly ignore the parameter name
  |
3 |     fn call(_: export_name);
  |             ~~~~~~~~~~~~~~

error[E0425]: cannot find function `a` in this scope
 --> /home/runner/work/glacier/glacier/ices/83471.rs:6:5
  |
6 |     a()
  |     ^ not found in this scope

error[E0658]: language items are subject to change
 --> /home/runner/work/glacier/glacier/ices/83471.rs:1:1
  |
1 | #[lang = "fn"]
  | ^^^^^^^^^^^^^^
  |
  = help: add `#![feature(lang_items)]` to the crate attributes to enable

error[E0601]: `main` function not found in crate `83471`
 --> /home/runner/work/glacier/glacier/ices/83471.rs:1:1
  |
1 | / #[lang = "fn"]
2 | | trait Fn {
3 | |     fn call(export_name);
4 | | }
5 | | fn call_through_fn_trait() {
6 | |     a()
7 | | }
  | |_^ consider adding a `main` function to `/home/runner/work/glacier/glacier/ices/83471.rs`

error[E0718]: `fn` language item must be applied to a trait with 1 generic argument
 --> /home/runner/work/glacier/glacier/ices/83471.rs:1:1
  |
1 | #[lang = "fn"]
  | ^^^^^^^^^^^^^^
2 | trait Fn {
  |         - this trait has 0 generic arguments

error: aborting due to 5 previous errors

Some errors have detailed explanations: E0425, E0601, E0658, E0718.
For more information about an error, try `rustc --explain E0425`.
==============