ices/105231.rs: fixed with errors
=== stdout ===
=== stderr ===
error[E0072]: recursive types `A` and `B` have infinite size
 --> /home/runner/work/glacier/glacier/ices/105231.rs:1:1
  |
1 | struct A<T>(B<T>);
  | ^^^^^^^^^^^ ---- recursive without indirection
2 | struct B<T>(A<A<T>>);
  | ^^^^^^^^^^^ ------- recursive without indirection
  |
help: insert some indirection (e.g., a `Box`, `Rc`, or `&`) to break the cycle
  |
1 ~ struct A<T>(Box<B<T>>);
2 ~ struct B<T>(Box<A<A<T>>>);
  |

error[E0275]: overflow evaluating the requirement `A<A<A<A<A<A<A<...>>>>>>>: Send`
  |
  = help: consider increasing the recursion limit by adding a `#![recursion_limit = "256"]` attribute to your crate (`105231`)
note: required because it appears within the type `B<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<A<u8>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>`
 --> /home/runner/work/glacier/glacier/ices/105231.rs:2:8
  |
2 | struct B<T>(A<A<T>>);
  |        ^

error: aborting due to 2 previous errors

Some errors have detailed explanations: E0072, E0275.
For more information about an error, try `rustc --explain E0072`.
==============